pyproject.toml
[project]
name = "online-store"
version = "0.1.0"
description = ""
authors = [
    {name = "Vadim Falaleev",email = "fvv14.12.14@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
requests = "^2.32.3"
mypy = "^1.15.0"
pytest-cov = "^6.0.0"


[tool.poetry.group.lint.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.1.2"

